feat: standardize output and add many fields
Includes a huge new amount of fields as per the tariff and also
updates our code to directly include a german email template.

Having the tempalte in code should help us closely manage future
change to it should we need to adopt it if the tariff gets updated.

There are also a bunch of fields that acr-cloud doesn't currently
deliver, but most of those are not in the "official" list as per
the tariff and only in the example.